execution/requirements.txt
anthropic
python-dotenv
requests
flask
flask-cors
gunicorn
google-generativeai>=0.8.3
